ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==11.0.0
numpy==2.1.0
scikit-learn==1.5.2
scikit-image==0.24.0
opencv-python-headless==4.10.0.84
scipy==1.14.1
joblib==1.4.2
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
pydantic==2.9.2
aiofiles==24.1.0
kagglehub==0.3.4
lightgbm==4.5.0
requests==2.32.3
python-dotenv==1.0.1
